+++ tests/e2e/write_test.go
@@ -24,3 +24,22 @@ func Test_E2E_Write(t *testing.T) {
 		StringMatches("Hello, world! Write was used."),
 	)
 }
+
+func Test_E2E_Write_Append(t *testing.T) {
+	RunTestWithNoConfigAndWithArgs(
+		t,
+		[]string{
+			"--route foo/bar",
+			fmt.Sprintf("--exec '%s'", strings.Join([]string{
+				`printf "Hello, world!" | {{MOCK_EXECUTABLE}} write`,
+				`printf " Append was used." | {{MOCK_EXECUTABLE}} write -a`,
+				`printf " Again." | {{MOCK_EXECUTABLE}} write --append`,
+			}, ";")),
+		},
+		"GET",
+		"foo/bar",
+		nil,
+		nil,
+		StringMatches("Hello, world! Append was used. Again."),
+	)
+}